DRY up one rspec test
Partially addresses #6.  Massively cleaner with better output.

The attribute get tests do not really deserve separate blocks.

The boilerplate bracketing is a waste of space, cognition and
maintenance energy. Everyone who has access to the tests has access to
the methods defined.  They don't need and shouldn't want a separate copy
of the method signature, let alone the entire method.  What they *do*
want is a better description of the behavior being tested, not just the
name of a method.

In particular, copying the entire implementation of the methods being
tested (for what, visual reference?) is fundamentally wrong.  The
implementation should be *free to change* without making the test
comments a liar. This spec already showed residue of drift in the
method names and descriptions.

Note: pulling out the boilerplate also revealed that one of the large
test blocks (harvest_directory) in fact is testing *nothing*!

I would like to get rid of the INSTANCE METHODS block level, but this is a
start.