Keep pry-byebug as a dev dependency even in CI
Without this change, pulling `main` and running `bundle install` winds up causing changes to `Gemfile.lock`; namely, adding `pry-byebug` and its dependencies. That is jarring and it is not clear that the reason we added this check five years ago is something we still care about.

The original issue being worked around is TravisCI failing to add pry-byebug to the bundle because it did not support Ruby 2.2: https://github.com/sul-dlss/moab-versioning/pull/149

I suspect the reason this dep is left out of the lockfile is that `ENV['CI']` is `true` when generating weekly dep update PRs on Jenkins.