Loosen validation of subdirectory names within content directory
Fixes #258

Now that SDR is beginning to allow users to upload content with path information, some paths within deposited items will contain directories named e.g. `data` or `content` or `metadata`. The moab-versioning gem flags these items as having forbidden content subdirectories. This is *not* part of the Moab specification, and we think this was some past attempt to prevent nested moab directory structures. We are now removing this consideration from the gem.